[BREAKING] Remove all exception handing from Request_Executor
Apps are now responsible for their own try/catch within their
index.php code. This allows for separate handling in shell tasks,
unit tests and the like and better separation of concerns : there's
no need to render a full HTML trace to the console if it's not in
web context, nor is a redirect or similar particularly helpful.