Do not throw an exception on requests to invalid URLS like //index.php?stuff=query
Various vulnerability probes and similar bots attempt to access urls that are
valid but unusual - e.g. //index.php?{some query string}. These generally
do not get interpreted the same way as normal URLs by apache etc, so drop
into Kohana's fallback URL handler. This should be stripping the querystring
from the text, consistent with the normal flow, but was not doing so.

Although these kinds of bot requests were being rejected, they were doing
so with a 500 and an error log entry, rather than a simple 404.

Fix the fallback handling so that it strips the querystring and works for
all URLs without errors.