Don't override session_cache_limiter when starting a session
As of 3.0.8 (commit 838385eb50fe2606302411663978eea645ec5e701)
kohana was calling `session_cache_limiter(FALSE)`. This prevented
php setting any default cache-control headers on responses
regardless of the configuration of `session.cache_limiter` in
php.ini. The behaviour was not optional, overridable, or
documented.

That commit was apparently added to fix a bug on the old kohana
issue tracker, which is no longer available.

We don't think this is a sane default - if handling the request
involved reading or writing $_SESSION then it is more likely
than not that the response should not be cached.

Applications that do want some or all responses to be cached
regardless of session state can still configure this either
by setting response headers in specific controllers or by
customising the php ini configuration of `session.cache_limiter`.