Fix race condition attempting to read cache files
We were very occasionally seeing an error `PHP Warning: ErrorException:
Automatic conversion of false to array is deprecated in
.../classes/Kohana/Core.php:746` in particular when a site is under
heavy load.

This was caused by `Kohana::cache('Kohana::find_file()')` returning
`false` (rather than `null` or `array`). That can only have been coming
from the result of the `unserialize(file_get_contents($dir.$file)` call.
The most likely cause of that would be `file_get_contents` returning
false due to a race condition where the file existed at the start of the
if block but no longer existing by the time we attempt to read the file.

My assumption is that with concurrent requests at the exact moment of
cache expiry it is possible for one request to delete the file just
before the other request tries to read it.

With this change we treat the failed read as any other failure to
unserialise the cache file and just return null. Additionally we cast
the null to `[]` specifically when loading the `find_files` cache so
that we can guarantee the `Kohana::$files` will always be an array.