!fix: further error case in `find_file` when cache read hits a race
The fix in #47 (released in 4.12.1) has reduced the frequency of this
error but has not eliminated it. There appears to be a remaining edge
case / race condition where the `Kohana::cache` method can return
FALSE rather than NULL.

This has proved difficult to replicate, the most likely explanation
appears to be that in certain conditions either the cache is being
serialized with an explicit `false` value, or that the `unserialize`
is somehow returning false without raising a PHP error (or at least that
this is not translated to an Exception).

Attempt to resolve by casting all falsey cache values to an empty array.

Additionally, strictly type the protected `Kohana_Core::$_files`
property as an array so that any remaining attempt to assign it as
`false` will trigger an error rather than this only being logged when
the code subsequently attempts to read it. This is technically a BC
break, but we do not expect any class to extend this property.