feat: Serve a generic 404 when route matches an abstract controller
Previously, this was throwing a generic exception, e.g. to alert
that a developer had potentially mis-configured routing.

However, this generates spurious alerts if:

a) a project still registers default routes e.g.
   `(<controller>(/<action>(/<id>)))`
b) a crawler requests a URL that is not intended to be defined, but
   happens to match an abstract class e.g. `/template/whatever` =>
   Controller_Template (even if the project itself does not use
   Controller_Template).

It is more logical to treat this as a generic 404, the same as if the
controller does not exist. User's functional testing should be
sufficient to know that all *expected* URLs e.g. from links generated by
the site are working as intended.